Small rewrite started.
